requirements.txt
asgiref==3.6.0
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.1.0
cryptography==40.0.2
Django==4.2
django-haystack==3.2.1
django-taggit==3.1.0
idna==3.4
isodate==0.6.1
Pillow==9.5.0
psycopg2-binary==2.9.6
pycparser==2.21
requests==2.28.2
six==1.16.0
sqlparse==0.4.4
typing_extensions==4.5.0
urllib3==1.26.15
Whoosh==2.7.4
